Removed v1, bump to v3, started refactoring
