+++ README.md
@@ -1,10 +1,9 @@
-[![GoDoc](https://godoc.org/github.com/lancer-kit/uwe/v2?status.png)](https://pkg.go.dev/github.com/lancer-kit/uwe/v2?tab=doc)
+[![GoDoc](https://godoc.org/github.com/lancer-kit/uwe/v3?status.png)](https://pkg.go.dev/github.com/lancer-kit/uwe/v3?tab=doc)
 [![Go Report Card](https://goreportcard.com/badge/github.com/lancer-kit/uwe)](https://goreportcard.com/report/github.com/lancer-kit/uwe)
 
 # uwe
 
-UWE (Ubiquitous Workers Engine) is a common toolset for building and organizing your Go application,  actor-like workers.  
-
+UWE (Ubiquitous Workers Engine) is a common toolset for building and organizing your Go application, actor-like workers.
 
 ## Table of Content
 
@@ -15,13 +14,12 @@ UWE (Ubiquitous Workers Engine) is a common toolset for building and organizing 
     2. [Worker](#worker)
     3. [Presets](#presets)
 
-
 ## Quick Start
 
 Get `uwe` using **go get**:
 
 ```shell script
-go get github.com/lancer-kit/uwe/v2
+go get github.com/lancer-kit/uwe/v3
 ```
 
 Here is an example HelloWorld service with HTTP API and background worker:
@@ -35,12 +33,11 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/lancer-kit/uwe/v2"
-	"github.com/lancer-kit/uwe/v2/presets/api"
+	"github.com/lancer-kit/uwe/v3"
+	"github.com/lancer-kit/uwe/v3/presets/api"
 )
 
-
-func main()  {
+func main() {
 	// fill configurations for the predefined worker that start an HTTP server
 	apiCfg := api.Config{
 		Host:              "0.0.0.0",
@@ -71,7 +68,7 @@ type dummy struct{}
 func NewDummy() uwe.Worker {
 	// At this point in most cases there we are preparing some state of the worker,
 	// like a logger, configuration, variable, and fields.
-	 return &dummy{}
+	return &dummy{}
 }
 
 // Init is an interface method used to initialize some state of the worker
@@ -114,19 +111,19 @@ func getRouter() http.Handler {
 
 ### Chief
 
-**Chief** is a supervisor that can be placed at the top of the go application's execution stack, 
-it is blocked until SIGTERM is intercepted and then it shutdown all workers gracefully.
-Also, `Chief` can be used as a child supervisor inside the `Worker`, which is launched by `Chief` at the top-level.
+**Chief** is a supervisor that can be placed at the top of the go application's execution stack, it is blocked until
+SIGTERM is intercepted and then it shutdown all workers gracefully. Also, `Chief` can be used as a child supervisor
+inside the `Worker`, which is launched by `Chief` at the top-level.
 
 ### Worker
 
 **Worker** is an interface for async workers which launches and manages by the **Chief**.
 
-1. `Init()` - method used to initialize some state of the worker that required interaction with outer context,
- for example, initialize some connectors. In many cases this method is optional, so it can be implemented as empty:
-  `func (*W) Init() error { return nil }`. 
-2. `Run(ctx Context) error` - starts the `Worker` instance execution. The context will provide a signal
- when a worker must stop through the `ctx.Done()`.
+1. `Init()` - method used to initialize some state of the worker that required interaction with outer context, for
+   example, initialize some connectors. In many cases this method is optional, so it can be implemented as empty:
+   `func (*W) Init() error { return nil }`.
+2. `Run(ctx Context) error` - starts the `Worker` instance execution. The context will provide a signal when a worker
+   must stop through the `ctx.Done()`.
 
 Workers lifecycle:
 
@@ -143,7 +140,8 @@ This library provides some working presets to simplify the use of `Chief` in pro
 
 #### HTTP Server
 
-`api.Server` is worker by default for starting a standard HTTP server. Server requires configuration and initialized `http.Handler`. 
+`api.Server` is worker by default for starting a standard HTTP server. Server requires configuration and
+initialized `http.Handler`.
 
 The HTTP server will work properly and will be correctly disconnected upon a signal from Supervisor (Chief).
 
@@ -157,7 +155,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/lancer-kit/uwe/v2"
+	"github.com/lancer-kit/uwe/v3"
 )
 
 func main() {
@@ -197,8 +195,8 @@ import (
 	"log"
 	"time"
 
-	"github.com/lancer-kit/uwe/v2"
-	"github.com/lancer-kit/uwe/v2/presets"
+	"github.com/lancer-kit/uwe/v3"
+	"github.com/lancer-kit/uwe/v3/presets"
 )
 
 func main() {
@@ -220,6 +218,7 @@ func main() {
 	chief.Run()
 }
 ```
+
 #### WorkerFunc
 
 `presets.WorkerFunc` is a type of worker that consist from one function. Allow to use the function as worker.
@@ -231,8 +230,8 @@ import (
 	"log"
 	"time"
 
-	"github.com/lancer-kit/uwe/v2"
-	"github.com/lancer-kit/uwe/v2/presets"
+	"github.com/lancer-kit/uwe/v3"
+	"github.com/lancer-kit/uwe/v3/presets"
 )
 
 func main() {
