+++ go.mod
@@ -1,12 +1,24 @@
-module github.com/lancer-kit/uwe
+module github.com/lancer-kit/uwe/v3
 
-go 1.12
+go 1.17
 
 require (
+	github.com/go-chi/chi v4.0.2+incompatible
 	github.com/go-ozzo/ozzo-validation/v4 v4.3.0
-	github.com/lancer-kit/sam v0.0.0-20200706130836-9a4a2be6c163
-	github.com/pkg/errors v0.9.1
-	github.com/sirupsen/logrus v1.8.1
-	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae // indirect
-	gopkg.in/yaml.v2 v2.2.4 // indirect
+	github.com/lancer-kit/sam v0.0.0-20190828205034-ab78e42fc7ce
+	github.com/pkg/errors v0.8.1
+	github.com/sirupsen/logrus v1.5.0
+	github.com/stretchr/testify v1.4.0
+	github.com/urfave/cli v1.22.2
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/konsorten/go-windows-terminal-sequences v1.0.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/russross/blackfriday/v2 v2.0.1 // indirect
+	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	golang.org/x/sys v0.0.0-20190422165155-953cdadca894 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
