+++ presets/job_test.go
@@ -4,7 +4,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/lancer-kit/uwe/v2"
+	"github.com/lancer-kit/uwe/v3"
 )
 
 func ExampleJob_Run() {
