+++ socket/entities.go
