--- v2/go.mod
@@ -1,15 +1,0 @@
-module github.com/lancer-kit/uwe/v2
-
-go 1.12
-
-require (
-	github.com/asaskevich/govalidator v0.0.0-20190424111038-f61b66f89f4a // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-chi/chi v4.0.2+incompatible
-	github.com/go-ozzo/ozzo-validation v3.6.0+incompatible
-	github.com/lancer-kit/sam v0.0.0-20190828205034-ab78e42fc7ce
-	github.com/pkg/errors v0.8.1
-	github.com/sirupsen/logrus v1.5.0
-	github.com/stretchr/testify v1.4.0
-	github.com/urfave/cli v1.22.2
-)
